ai/moderation-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
transformers==4.37.0
torch==2.1.2
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
httpx==0.26.0
python-dotenv==1.0.0
redis==5.0.1
spacy==3.7.2
